Organize components into subdirectories
