+++ frontend/src/components/Home/Home.tsx
@@ -6,12 +6,12 @@ import { DynamicMix } from '../../models/DynamicMix';
 import { SongData } from '../../models/SongData';
 import { StaticMix } from '../../models/StaticMix';
 import HomeNavBar from '../Nav/HomeNavBar';
-import DeleteDynamicMixModal from '../SongTable/DeleteDynamicMixModal';
-import DeleteStaticMixModal from '../SongTable/DeleteStaticMixModal';
-import DeleteTrackModal from '../SongTable/DeleteTrackModal';
-import DynamicMixModal from '../SongTable/DynamicMixModal';
+import DeleteDynamicMixModal from '../SongTable/Modal/DeleteDynamicMixModal';
+import DeleteStaticMixModal from '../SongTable/Modal/DeleteStaticMixModal';
+import DeleteTrackModal from '../SongTable/Modal/DeleteTrackModal';
+import DynamicMixModal from '../SongTable/Modal/DynamicMixModal';
+import StaticMixModal from '../SongTable/Modal/StaticMixModal';
 import SongTable from '../SongTable/SongTable';
-import StaticMixModal from '../SongTable/StaticMixModal';
 import UploadModal from '../Upload/UploadModal';
 import './Home.css';
 import MusicPlayer from './MusicPlayer';
