+++ frontend/src/components/SongTable/Button/DeleteStaticMixButton.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { Button } from 'react-bootstrap';
 import { Trash } from 'react-bootstrap-icons';
-import { StaticMix } from '../../models/StaticMix';
+import { StaticMix } from '../../../models/StaticMix';
 
 interface Props {
   disabled?: boolean;
