+++ frontend/src/components/SongTable/Button/PausePlayButton.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { Button } from 'react-bootstrap';
 import { PauseFill, PlayFill } from 'react-bootstrap-icons';
-import { SongData } from '../../models/SongData';
-import { StaticMix } from '../../models/StaticMix';
+import { SongData } from '../../../models/SongData';
+import { StaticMix } from '../../../models/StaticMix';
 
 interface Props {
   song: SongData | StaticMix;
@@ -13,6 +13,9 @@ interface Props {
   onPlayClick: (song: SongData | StaticMix) => void;
 }
 
+/**
+ * Component for pause/play button shown in the song table.
+ */
 class PausePlayButton extends React.Component<Props> {
   handlePlay = (): void => {
     if (this.props.playing) {
@@ -35,6 +38,7 @@ class PausePlayButton extends React.Component<Props> {
         {playing && !disabled ? (
           <PauseFill size={28} />
         ) : (
+          // Play icon is slightly off-centre when placed in circular button
           <PlayFill size={28} style={{ marginLeft: 1.25, marginRight: -1.25 }} />
         )}
       </Button>
