+++ frontend/src/components/SongTable/Modal/DynamicMixModal.tsx
@@ -1,44 +1,34 @@
 import axios from 'axios';
 import * as React from 'react';
 import { Button, Modal } from 'react-bootstrap';
-import { SongData } from '../../models/SongData';
-import { StaticMix } from '../../models/StaticMix';
-import StaticMixModalForm from './Form/StaticMixModalForm';
+import { DynamicMix } from '../../../models/DynamicMix';
+import { SongData } from '../../../models/SongData';
+import DynamicMixModalForm from '../Form/DynamicMixModalForm';
 
 interface Props {
   song?: SongData;
   show: boolean;
-  hide: () => void;
   exit: () => void;
+  hide: () => void;
   refresh: () => void;
-  submit: (srcId: string, id: string, status: string) => void;
+  submit: (id: string) => void;
 }
 
 interface State {
   model: string;
   randomShifts: number;
-  vocals: boolean;
-  drums: boolean;
-  bass: boolean;
-  other: boolean;
-  overwrite: boolean;
   errors: string[];
 }
 
 /**
- * Component for the source separation modal.
+ * Modal for creating dynamic mix.
  */
-class StaticMixModal extends React.Component<Props, State> {
+class DynamicMixModal extends React.Component<Props, State> {
   constructor(props: Props) {
     super(props);
     this.state = {
       model: 'spleeter',
       randomShifts: 0,
-      vocals: false, // Include vocals
-      drums: false, // Include drums
-      bass: false, // Include bass
-      other: false, // Include accompaniment
-      overwrite: false, // Whether to overwrite existing static mix, if exists
       errors: [],
     };
   }
@@ -50,11 +40,6 @@ class StaticMixModal extends React.Component<Props, State> {
     this.setState({
       model: 'spleeter',
       randomShifts: 0,
-      vocals: false,
-      drums: false,
-      bass: false,
-      other: false,
-      overwrite: false,
       errors: [],
     });
   };
@@ -87,19 +72,15 @@ class StaticMixModal extends React.Component<Props, State> {
       source_track: this.props.song.id,
       separator: this.state.model,
       random_shifts: this.state.randomShifts,
-      vocals: this.state.vocals,
-      drums: this.state.drums,
-      bass: this.state.bass,
-      other: this.state.other,
-      overwrite: this.state.overwrite,
+      overwrite: true,
     };
 
-    // Make request to add Song
+    // Make API request to create the mix
     axios
-      .post<StaticMix>('/api/mix/static/', data)
+      .post<DynamicMix>('/api/mix/dynamic/', data)
       .then(({ data }) => {
         this.props.hide();
-        this.props.submit(data.source_track, data.id, data.status);
+        this.props.submit(data.id);
       })
       .catch(({ response }) => {
         const { data } = response;
@@ -107,11 +88,6 @@ class StaticMixModal extends React.Component<Props, State> {
           errors: data.errors,
         });
       });
-  };
-
-  handleCheckboxChange = (event: React.ChangeEvent<HTMLInputElement>): void => {
-    const { name, checked } = event.target;
-    this.setState({ [name]: checked, errors: [] } as any);
   };
 
   handleModelChange = (event: React.ChangeEvent<HTMLSelectElement>): void => {
@@ -127,28 +103,21 @@ class StaticMixModal extends React.Component<Props, State> {
   };
 
   render(): JSX.Element | null {
-    const { vocals, drums, bass, other, errors } = this.state;
+    const { errors } = this.state;
     const { show, song } = this.props;
     if (!song) {
       return null;
     }
 
-    // Display error if all or no parts are checked
-    const allChecked = vocals && drums && bass && other;
-    const noneChecked = !(vocals || drums || bass || other);
-
     return (
       <Modal show={show} onHide={this.onHide} onExited={this.onExited}>
         <Modal.Header closeButton>
-          <Modal.Title>Create static mix</Modal.Title>
+          <Modal.Title>Create dynamic mix</Modal.Title>
         </Modal.Header>
         <Modal.Body>
-          <StaticMixModalForm
+          <DynamicMixModalForm
             song={song}
-            allChecked={allChecked}
-            noneChecked={noneChecked}
             errors={errors}
-            handleCheckboxChange={this.handleCheckboxChange}
             handleModelChange={this.handleModelChange}
             handleRandomShiftsChange={this.handleRandomShiftsChange}
           />
@@ -157,7 +126,7 @@ class StaticMixModal extends React.Component<Props, State> {
           <Button variant="outline-danger" onClick={this.onHide}>
             Cancel
           </Button>
-          <Button variant="primary" disabled={allChecked || noneChecked} onClick={this.onSubmit}>
+          <Button variant="success" onClick={this.onSubmit}>
             Create Mix
           </Button>
         </Modal.Footer>
@@ -166,4 +135,4 @@ class StaticMixModal extends React.Component<Props, State> {
   }
 }
 
-export default StaticMixModal;
+export default DynamicMixModal;
